ajout de aircrafts les avions

+++ app/controllers/api/v1/aircrafts_controller.rb
@@ -9,8 +9,12 @@ module Api
 
       # GET /api/v1/aircrafts/:id
       def show
-        aircraft = Aircraft.find(params[:id])
-        render json: aircraft
+        aircraft = Aircraft.find_by(id: params[:id])
+        if aircraft
+          render json: aircraft
+        else
+          render json: { error: "Aircraft not found" }, status: :not_found
+        end
       end
 
       # POST /api/v1/aircrafts
@@ -25,29 +29,49 @@ module Api
 
       # PUT /api/v1/aircrafts/:id
       def update
-        aircraft = Aircraft.find(params[:id])
-        if aircraft.update(aircraft_params)
-          render json: aircraft
+        aircraft = Aircraft.find_by(id: params[:id])
+        if aircraft
+          if aircraft.update(aircraft_params)
+            render json: aircraft
+          else
+            render json: { errors: aircraft.errors.full_messages }, status: :unprocessable_entity
+          end
         else
-          render json: { errors: aircraft.errors.full_messages }, status: :unprocessable_entity
+          render json: { error: "Aircraft not found" }, status: :not_found
         end
       end
 
       # DELETE /api/v1/aircrafts/:id
       def destroy
-        aircraft = Aircraft.find(params[:id])
-        aircraft.destroy
-        head :no_content
+        aircraft = Aircraft.find_by(id: params[:id])
+        if aircraft
+          aircraft.destroy
+          head :no_content
+        else
+          render json: { error: "Aircraft not found" }, status: :not_found
+        end
+      end
+
+      # POST /api/v1/aircrafts/import
+      def import
+        # Vérifie si un fichier a bien été soumis
+        if params[:file].present?
+          Aircraft.import_csv(params[:file])  # Appelle la méthode d'importation
+          render json: { message: "Data imported successfully" }, status: :ok
+        else
+          render json: { error: "No file uploaded" }, status: :unprocessable_entity
+        end
       end
 
       private
 
       def aircraft_params
+        # Permet de spécifier les paramètres autorisés pour les avions
         params.require(:aircraft).permit(
-          :icao24, :acars, :adsb, :built, :categoryDescription, :country, :engines, 
-          :firstFlightDate, :firstSeen, :icaoAircraftClass, :lineNumber, :manufacturerIcao, 
-          :manufacturerName, :model, :modes, :nextReg, :notes, :operator, :operatorCallsign, 
-          :operatorIata, :operatorIcao, :owner, :prevReg, :regUntil, :registered, :registration, 
+          :icao24, :timestamp, :acars, :adsb, :built, :categoryDescription, :country, :engines,
+          :firstFlightDate, :firstSeen, :icaoAircraftClass, :lineNumber, :manufacturerIcao,
+          :manufacturerName, :model, :modes, :nextReg, :notes, :operator, :operatorCallsign,
+          :operatorIata, :operatorIcao, :owner, :prevReg, :regUntil, :registered, :registration,
           :selCal, :serialNumber, :status, :typecode, :vdl
         )
       end
